Level update
With the updated Object Oriented API, this commit includes the
clearBlockSet() method to execute the OO command blockSet.clear().
Clearing the block set is required before new blocks can be properly
added into a level.
This commit also includes the LevelDemo file, heavily based on NinDemo.
This demonstrates using the procedural API to change the level.